app/validator/requirements.txt
argparse==1.4.0
pyyaml==6.0.1
# Add other dependencies as needed for your validator or app
